[net processing] Add peerman_impl.h
